fix: resolve all clippy warnings that were failing CI
- config.rs: match -> if let for single-pattern match
- debug_log.rs: remove needless return statement
- game.rs: remove let binding before return (let_and_return)
- text.rs: remove redundant u32 -> u32 casts in set_pixel call

+++ src/text.rs
@@ -152,7 +152,7 @@ impl BitmapFont {
                 let y = oy + py as u32;
                 for px in 0..glyph_w_bits {
                     if (row_bits & (1 << px)) != 0 {
-                        image.set_pixel((ox + px) as u32, y as u32, WHITE);
+                        image.set_pixel(ox + px, y, WHITE);
                     }
                 }
             }
